modify variable name

+++ action.sh
@@ -6,6 +6,12 @@ set -o pipefail
 # Populate defaults
 [[ -n $GITHUB_ACTION_PATH ]] || GITHUB_ACTION_PATH=$(pwd)
 
-echo "ℹ️ Copying semgrep file to repo root directory"
-cp "$GITHUB_ACTION_PATH"/semgrep/${INTEGRATION}.yml .semgrep.yml
-cp "$GITHUB_ACTION_PATH"/semgrep/.semgrepignore .semgrepignore
+if [[ -f "$GITHUB_ACTION_PATH"/semgrep/${REPOSITORY}.yml ]]
+then
+    echo "ℹ️ Copying semgrep files to repo root directory"
+    cp "$GITHUB_ACTION_PATH"/semgrep/${REPOSITORY}.yml .semgrep.yml
+    cp "$GITHUB_ACTION_PATH"/semgrep/.semgrepignore .semgrepignore
+else
+  echo "❌ Semgrep file does not exist for REPOSITORY '${REPOSITORY}'"
+  return 1
+fi
